Split app out into Flutter recommended structure. Fixed button on perk configuration page. Fixed styling issues on Drawer
